SIB-138 SIB -141
fixing CoinController tests

In last commit on brunch 138 hibernate-validator was removed from pom It was needed, so I added back
 cleaning up Error handling
 I moved handleConstraintViolationException  to ErrorHandlingAdvice
