+++ pom.xml
@@ -61,6 +61,10 @@
     <!-- End of Spring dependencies -->
 
     <!-- Non-Spring dependencies -->
+    <dependency>
+      <groupId>org.hibernate.validator</groupId>
+      <artifactId>hibernate-validator</artifactId>
+    </dependency>
 
     <dependency>
       <groupId>com.github.ben-manes.caffeine</groupId>
@@ -155,6 +159,7 @@
             <configuration>
               <excludes>
                 <exclude>com/capgemini/fs/coindashboard/controller/utils/*</exclude>
+                <exclude>com/capgemini/fs/coindashboard/CoinDashboardApplication.*</exclude>
               </excludes>
               <rules>
                 <rule>
