+++ src/main/java/com/capgemini/fs/coindashboard/controller/CoinController.java
@@ -6,7 +6,6 @@ import com.capgemini.fs.coindashboard.cacheService.CacheService;
 import com.capgemini.fs.coindashboard.controller.exceptionHandler.CoinNotFoundException;
 import com.capgemini.fs.coindashboard.controller.exceptionHandler.ServerIsNotRespondingException;
 import com.capgemini.fs.coindashboard.controller.utils.ValidTimestamp;
-import javax.validation.ConstraintViolationException;
 import javax.validation.constraints.Max;
 import javax.validation.constraints.Min;
 import javax.validation.constraints.NotBlank;
@@ -14,15 +13,12 @@ import javax.validation.constraints.Size;
 import lombok.AllArgsConstructor;
 import org.springframework.beans.factory.annotation.Autowired;
 import org.springframework.http.HttpHeaders;
-import org.springframework.http.HttpStatus;
 import org.springframework.http.ResponseEntity;
 import org.springframework.validation.annotation.Validated;
-import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.GetMapping;
 import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestParam;
-import org.springframework.web.bind.annotation.ResponseStatus;
 import org.springframework.web.bind.annotation.RestController;
 import org.springframework.web.servlet.support.ServletUriComponentsBuilder;
 
@@ -105,12 +101,5 @@ public class CoinController {
             .build()
             .toUriString();
     return ResponseEntity.status(OK).header(HttpHeaders.LOCATION, location).body(chart);
-  }
-
-  @ExceptionHandler(ConstraintViolationException.class)
-  @ResponseStatus(HttpStatus.BAD_REQUEST)
-  ResponseEntity<String> handleConstraintViolationException(ConstraintViolationException e) {
-    return new ResponseEntity<>(
-        "not valid due to validation error: " + e.getMessage(), HttpStatus.BAD_REQUEST);
   }
 }
