+++ src/main/java/com/capgemini/fs/coindashboard/controller/exceptionHandler/ErrorHandlingAdvice.java
@@ -1,6 +1,8 @@
 package com.capgemini.fs.coindashboard.controller.exceptionHandler;
 
+import javax.validation.ConstraintViolationException;
 import org.springframework.http.HttpStatus;
+import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.ControllerAdvice;
 import org.springframework.web.bind.annotation.ExceptionHandler;
 import org.springframework.web.bind.annotation.ResponseBody;
@@ -30,5 +32,10 @@ public class ErrorHandlingAdvice {
     return ex.getMessage();
   }
 
-
+  @ExceptionHandler(ConstraintViolationException.class)
+  @ResponseStatus(HttpStatus.BAD_REQUEST)
+  ResponseEntity<String> handleConstraintViolationException(ConstraintViolationException e) {
+    return new ResponseEntity<>(
+        "not valid due to validation error: " + e.getMessage(), HttpStatus.BAD_REQUEST);
+  }
 }
