+++ src/test/java/com/capgemini/fs/coindashboard/controller/CoinControllerChartTest.java
@@ -8,23 +8,30 @@ import static org.springframework.test.web.servlet.result.MockMvcResultMatchers.
 
 import com.capgemini.fs.coindashboard.cacheService.CacheService;
 import java.util.Optional;
+import org.junit.jupiter.api.BeforeEach;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
+import org.springframework.test.web.servlet.setup.MockMvcBuilders;
+import org.springframework.web.context.WebApplicationContext;
 
-
-@AutoConfigureMockMvc
 @ExtendWith(SpringExtension.class)
-@WebMvcTest
+@WebMvcTest(controllers = CoinController.class)
 class CoinControllerChartTest {
 
-  @Autowired private MockMvc mockMvc;
   @MockBean private CacheService cacheService;
+  @Autowired private WebApplicationContext context;
+
+  private MockMvc mockMvc;
+
+  @BeforeEach
+  public void setup() {
+    mockMvc = MockMvcBuilders.webAppContextSetup(context).build();
+  }
 
   @Test
   public void getAllThrowsError400whenChart_toBeforeChart_from() throws Exception {
