+++ src/test/java/com/capgemini/fs/coindashboard/controller/CoinControllerMarketDataTest.java
@@ -11,16 +11,13 @@ import java.util.Optional;
 import org.junit.jupiter.api.Test;
 import org.junit.jupiter.api.extension.ExtendWith;
 import org.springframework.beans.factory.annotation.Autowired;
-import org.springframework.boot.test.autoconfigure.web.servlet.AutoConfigureMockMvc;
 import org.springframework.boot.test.autoconfigure.web.servlet.WebMvcTest;
 import org.springframework.boot.test.mock.mockito.MockBean;
 import org.springframework.test.context.junit.jupiter.SpringExtension;
 import org.springframework.test.web.servlet.MockMvc;
 
-
 @ExtendWith(SpringExtension.class)
-@AutoConfigureMockMvc
-@WebMvcTest
+@WebMvcTest(controllers = CoinController.class)
 class CoinControllerMarketDataTest {
 
   @Autowired private MockMvc mockMvc;
